# Faster knotvector addition