This PR fixes #275.