@hyrodium 2022-09-10
**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools, Random

julia> Random.seed!(42)
TaskLocalRNG()

julia> k1, k2 = KnotVector(rand(20)), KnotVector(rand(30));

julia> @benchmark $k1+$k2
BenchmarkTools.Trial: 10000 samples with 218 evaluations.
 Range (min … max):  327.959 ns …   3.355 μs  ┊ GC (min … max): 0.00% … 85.04%
 Time  (median):     346.528 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   359.556 ns ± 110.967 ns  ┊ GC (mean ± σ):  1.39% ±  4.08%

   ▃▄▇████▇▅▃▁▁    ▁▁ ▁                           ▁             ▂
  ▇█████████████████████▇▆▇▅▅▆▆▅▅▅▅▅▆▄▅▃▄▅▅▂▂▅▅▅▇▇█████▇▇▇▇▆▇██ █
  328 ns        Histogram: log(frequency) by time        505 ns <

 Memory estimate: 496 bytes, allocs estimate: 1.

julia> k1, k2 = KnotVector(rand(2)), KnotVector(rand(3));

julia> @benchmark $k1+$k2
BenchmarkTools.Trial: 10000 samples with 982 evaluations.
 Range (min … max):  59.879 ns …  1.286 μs  ┊ GC (min … max): 0.00% … 91.15%
 Time  (median):     64.995 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   70.480 ns ± 42.128 ns  ┊ GC (mean ± σ):  2.55% ±  4.14%

  ▁▂  ▆█▆▂▁▁      ▁▂▁                                         ▁
  ███████████▇█▇▆▇███▇▇▆▅▅▅▄▅▅▅▆████▇▇▇███▇▅▆▅▄▅▆▅▆▇▇▇▇▅▆▆▆▆▄ █
  59.9 ns      Histogram: log(frequency) by time       114 ns <

 Memory estimate: 96 bytes, allocs estimate: 1.

julia> @benchmark 3*$k2
BenchmarkTools.Trial: 10000 samples with 870 evaluations.
 Range (min … max):  134.554 ns …  1.543 μs  ┊ GC (min … max): 0.00% … 89.31%
 Time  (median):     147.509 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   156.777 ns ± 77.583 ns  ┊ GC (mean ± σ):  3.22% ±  5.88%

  ▂▃▂ ▄▇█▅▄▂▁▁   ▁▃▁                                           ▁
  ████████████▇▇█████▇▆▆▆▆▅▆▅▅▄▄▅▅▅▅▃▄▆▆▆▇███▆▅▅▅▃▃▄▃▃▃▅▅▅▄▄▃▃ █
  135 ns        Histogram: log(frequency) by time       258 ns <

 Memory estimate: 240 bytes, allocs estimate: 2.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools, Random

julia> Random.seed!(42)
TaskLocalRNG()

julia> k1, k2 = KnotVector(rand(20)), KnotVector(rand(30));

julia> @benchmark $k1+$k2
BenchmarkTools.Trial: 10000 samples with 898 evaluations.
 Range (min … max):  124.210 ns …  1.149 μs  ┊ GC (min … max): 0.00% … 80.91%
 Time  (median):     206.371 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   203.975 ns ± 59.180 ns  ┊ GC (mean ± σ):  2.51% ±  7.07%

                             █▇                                 
  ▁▁▁▂▅▄▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▃▄▆██▇▄▆▄▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  124 ns          Histogram: frequency by time          301 ns <

 Memory estimate: 496 bytes, allocs estimate: 1.

julia> k1, k2 = KnotVector(rand(2)), KnotVector(rand(3));

julia> @benchmark $k1+$k2
BenchmarkTools.Trial: 10000 samples with 990 evaluations.
 Range (min … max):  42.727 ns …  1.052 μs  ┊ GC (min … max): 0.00% … 95.50%
 Time  (median):     50.151 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   52.367 ns ± 38.550 ns  ┊ GC (mean ± σ):  3.24% ±  4.21%

            ▂▇█▃                                               
  ▂▃▇▇▅▅▄▄▄▅████▄▃▂▂▂▂▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  42.7 ns         Histogram: frequency by time        80.9 ns <

 Memory estimate: 96 bytes, allocs estimate: 1.

julia> @benchmark 3*$k2
BenchmarkTools.Trial: 10000 samples with 896 evaluations.
 Range (min … max):  126.914 ns …  1.061 μs  ┊ GC (min … max): 0.00% … 87.57%
 Time  (median):     129.542 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   134.138 ns ± 38.559 ns  ┊ GC (mean ± σ):  1.36% ±  4.16%

  ▁▆██▇▃▁▁▁▁▁            ▁▁▁                                   ▂
  ████████████▇▆▆▆▆▆▆▆▄▆▅███▇▆▄▆▆▅▆▅▅▄▄▅▆▆▇███▆▆▆▇▆▇███▇▆▆▆▆▅▅ █
  127 ns        Histogram: log(frequency) by time       173 ns <

 Memory estimate: 128 bytes, allocs estimate: 1.
```