@hyrodium 2022-09-10
This PR also fixes the type-instability with `*(::Int, ::UniformKnotVector)`

**Before this PR**
```julia
julia> using BasicBSpline

julia> k = UniformKnotVector(1:15)
UniformKnotVector(1:15)

julia> 0*k
EmptyKnotVector{Int64}()

julia> 2*k
KnotVector([1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15])
```

**After this PR**

```julia
julia> using BasicBSpline

julia> k = UniformKnotVector(1:15)
UniformKnotVector(1:15)

julia> 0*k
KnotVector(Int64[])

julia> 2*k
KnotVector([1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15])
```